add 'node_type' to list of immutable model fields
In the conversion from a backend entity to an ORM entity, the
`node_type` attribute of the ModelWrapper is requested, which triggered
a database query.
This meant that a QueryBuilder returning N nodes actually caused 1+N
queries to the database instead of 1 (with obvious performance hits).

By making 'node_type' explicitly immutable (changing the type of a node
should not be allowed), requesting it no longer triggers a database
query.