Add 'verdi node repo cp' command.
The command takes a node pk / uuid, zero or more file names, and an
ouput directory as arguments. By default, it will create a sub-directory
with the node uuid in the output directory, and copy the files with
the given names from the repository to that directory. If no file names
are given, all objects in the repository are copied.
The command has two optional flags:
 --no-uuid: copies the files directly into the output directory, without
            uuid prefix
 --force: Copy files even if they already exist at the target destination