Add test for raising an error when a broken workchain is documented.
Adds a test of the sphinx extension where a workchain that raises
an error in its 'define' method is documented. Due to plumpy PR#136,
the error should propagate into the sphinx build.
On the side of the AiiDA sphinx extension, we catch the error and
re-raise with a more helpful error message.