Change name and options in 'file copying' command.
The command is renamed to 'verdi node repo dump'. The option
of specifying specific file names to be copied is removed. The
output paths no longer prefixed with a UUID. When a file or directory
exists already, the command prompts to either abort, skip the current
file / directory, overwrite the file / directory, or (in the case
of directories only) merge the contents. A 'force' option can be
given to always overwrite files / directories.
Tests are not yet updated to the changed interface, since we want
to get feedback on that first.