Make sure that datetime conversions ignore `None`
The functions `datetime_to_isoformat` and `isoformat_to_datetime`
assumed to always have a proper type (string or datetime) as input.
However, in some cases, they were called with values that could
be potentially `None`, like this in aiida/engine/utils.py:
```
timezone.isoformat_to_datetime(manager.get(key).value))
```
We are now directly returning `None` if `None` is passed as an
input (and then it's up to the caller to then decide what to do
with the value.

Fixes #3336