Add option to expand namespaces in sphinx directive.
Adds the :expand-namespaces: flag to the sphinx directive, which
causes the <details> tags to be created with 'open="open"'
attributes. This makes them expand by default (collapsible by
clicking). Added this case to the test / demo documentation of the
sphinx extension, and in the documentation for `CalcJob`, where
the options are explained.