Fix two issues in `verdi computer configure ssh`
The interactive SSH was not accepting an empty
ssh key_filename, while this should be acceptable.
I am also adding a test for the interactive setup of
SSH computers.

Moreover, when reconfiguring a computer, the value
of the cooldown time wasn't resued from the authinfo;
instead, the class default was also reused due to some
hardcoding. Code now behaves correctly and is simplified
avoid duplication in base classes and subclasses.

Fixes #3633 and fixes #3634