Add hook and exit_code keyword for more nuanced cache validation.
Adds a method `is_valid_cache` to the `Process` class, which is
called with a `ProcessNode` instance as an argument. This method
is called by `ProcessNode.is_valid_cache`. It allows implementing
custom behavior in `Process` sub-classes, such as the `CalcJob`
classes implemented in plugins.

As a default, the `Process.is_valid_cache` checks whether the
exit code returned from the process has a -- newly introduced --
`invalidates_cache` attribute set to `True`. This is a simple
way of marking the as cache invalid for a specific exit code only.

Note that this is a backwards-incompatible change, which can
break code that uses unpacking of ExitCode, for example

    status, message = ExitCode(...)

The new ways of controlling caching are documented in the
developer's documentation.