Add more methods to control cache invalidation of completed process node (#3637)
A new attribute `invalidates_cache` is added to the `ExitCode` named
tuple which is false by default. If a process node has an exit status
that corresponds to an exit code with `invalidate_cache=False`, the node
should not be considered for caching.

This functionality is leveraged in the `ProcessNode.is_valid_cache`
method which will return false if the exit code set on the node belongs
to an invalid cache exit code. This method provides an additional hook
over the same method on the `ProcessNode` class to allow plugins to add
additional custom logic in `Process` sub-classes.

Note that this is a backwards-incompatible change, which can break code
that uses unpacking of `ExitCode` tuples, for example:

    status, message = ExitCode(...)

Finally, the new `Process.is_valida_cache` implementation requires the
class to be importable for it to be considered cachable. In the case of
process functions this means that the function is importable. This meant
that process function definitions in the unit tests had to be moved to
the module level.