Updated manifest to contain correct backup tmpl
This was moved to manage/backup/backup_info.json.tmpl but the MANIFEST
was never updated meaning that anyone pip installing aiida would be
missing this file and creating a backup according to our instructions
[1] would fail.

I also checked the other files in the manifest and they're fine.
Also checked that any references to the code are correct.

[1] https://aiida.readthedocs.io/projects/aiida-core/en/latest/backup/index.html