Add the `BaseRestartWorkChain`
The `BaseRestartWorkChain` is designed to help the writing of a base
workchain that wraps the launching of a sub process, for example a
calculation job, and provides a framework for easily adding automated
error handling and performing sanity checks.

The class and its utilities wer originally implemented in the
`aiida-quantumespresso` plugin but was designed from the get-go to be
generically applicable. It was quickly adopted and used in other
plugins. Therefore we are now moving it to `aiida-core` so that it can
be useful to many plugins without separate copies of the code having to
be maintained. Originally it was only designed to wrap around
calculation jobs, but the concept is generic enough that here it is
generalized to any sub process, such as `WorkChains`.

The basic concept is simple, instead of subclassing from `WorkChain` one
subclasses the `BaseRestartWorkChain` and use at a minimum the outline:

   cls.setup
   while_(cls.should_run_process)(
       cls.run_process,
       cls.inspect_process,
   ),
   cls.results

The logic of those outline methods is implemented on the base class so
the only thing that remains is to specify the process sub class that
needs to be used. The sub process will be launched until it is sucessful,
i.e. exit status 0, or the maximum number of iterations is exceeded. The
`inspect_process` will loop over the list of registered process handlers.
The handlers can check for errors in case that the sub process failed, or
perform sanity checks even if the sub process was successful. The
handlers can return a `ProcessHandlerReport` to control the further flow,
for example by breaking out of the process handler call loop or even
completely aborting the workchain if an unrecoverable problem was
encountered.