Add test that Sphinx process extension raises when process spec raises (#3617)
A problem in the implementation of the process spec construction in the
`plumpy` library resulted in exceptions being raised during the `spec()`
call being swallowed and an incomplete spec being set. As a result the
compiled process documentation would be missing parts of the spec with
no indication of what went wrong.

The `Process.spec()` method was improved to not set the spec and reraise
if any exceptions are raised. This allows the Sphinx extension to catch
the error and re-raise with a more helpful error message.