Make command line deprecation warnings visible with test profile
The `deprecated_command` decorator was only printing the warning if the
profile was not a test profile. With `verdi devel tests` being
deprecated, typically run with a test profile, the warning was being
swallowed.

This change gave problems with a test of `verdi comment show` that
expected no output for no input, but now of course failed due to the
output of the deprecation. Since it is a deprecated command anyway I
simply remove the test.