Move `CalcJob.presubmit` call from `CalcJob.run` to `Waiting.execute` (#3666)
The `presubmit` calls through to `prepare_for_submission` which is the
place where the `CalcJob` implementation writes the input files based on
the inputs nodes in a temporary sandbox folder. The contents of which
are then stored in the repository of the node before the process
transitions to the waiting state in await of the upload task. This
upload task will then copy over the contents of the node repository to
the working directory on the remote computer.

There is a use case to limit which files are copied to the node's
repository from the sandbox folder used by `prepare_for_submission`.
This means that the sandbox folder will have to be passed to the
`upload_calculation`. However, since the creation of this happens in the
`CalcJob.run` call, which is quite far removed from the eventual use,
this opens it up for a lot of problems. The time between folder
population by `prepare_for_submission` and eventual use in the upload
task can be significant, among other things because the upload task will
have to wait for transport. To make this feasible the creation of the
sandbox folder has to be moved closer to the upload task.

Here we move the `presubmit` call from `CalcJob.run` to the upload
transport task which will create the sandbox folder and pass it into the
`upload_calculation`. This limits the time frame in which there is a
chance for the contents of the sandbox to get lost. In addition, this
now gives additional freedom of deciding which files from the sandbox
are permanently stored in the node's repository for extra provenance.