Update `pyyaml` to prevent arbitrary code execution  (#3675)
* Fix various deprecation warnings

 * Do not use `Test` as prefix for dummy classes
 * Replace `imp` for `importlib` in REST API
 * Use `identifier` instead of deprecated `node_class`

* Update `pyyaml` to prevent arbitrary code execution

Before `pyyaml==5.1` the `yaml.load` function was vulnerable to
arbitrary code execution, because it loaded the full set of YAML. There
was an alternative `safe_load` but this was not the default and could
only load a sub set of the markup language. The new version of pyyaml
deprecates the old vulnerable code and provides the `FullLoader` that
can load the full set without being vulnerable.