fix performance issue when exporting many groups
When providing groups to `verdi export`, it was looping over all groups
and using the `Group.nodes` iterator to retrieve the nodes contained in
the groups. This results (at least) in one query per group, and is
therefore every inefficient for large numbers of groups. The new
implementation replaces this by two queries, one for Data nodes and one
for Process nodes.
It also no longer constructs the ORM objects since they are unnecessary.