Deal with unreachable daemon worker in `get_daemon_status`
The `aiida.cmdline.utils.daemon.get_daemon_status` utility function used
in `verdi daemon status` calls the `DaemonClient.get_worker_info` method
which will ask the circus client to get information of the daemon
workers it is managing. Under normal conditions this will return a
dictionary of worker pids with a dictionary of their stats. However,
sometimes the daemon may fail to retrieve these stats and the dictionary
is replaced with a string containing an error message. The
`get_daemon_status` method now deals with this elegantly and temporary
prints `-` as placeholder for the unknown statistics. Since this is
typically caused by a transient problem the next time the command is
called the correct information will be displayed.