Add feature AGE
The AiiDA Graph Explorer (AGE) is a general purpose tool to perform
graph traversal of AiiDA graphs. It considers AiiDA nodes and groups
(eventually even computers and users) as if they were both 'graph nodes'
of an 'expanded graph', and generalizes the exploration of said graph.
The 'rules' that indicate how to traverse this graph are configured by
using generic querybuilder instances (i.e. with information about the
connections but without specific initial nodes/groups and without any
projections). The initial set of nodes/groups is provided directly to
the rule, which then will perform successive applications of the query,
each on top of the results of the previous one. This cycle is repeated
for a specified number of time, which can be specified to be 'until no
new nodes are found'.

The current implementation works with the following (public) classes:

* Basket: generic container class that can store sets of nodes, groups,
node-node edges (aiida links) and group-node edges. These are the
objects that the rule-objects receive and return.

* UpdateRule: initialized with a querybuilder instance (and optionally a
max number of iterations and the option to track edges), it can then be
run with an initial set of nodes to obtain the result of the accumulated
traversal procedure described by the iterations of the query.

* ReplaceRule: same as the update rule, except that at the end of the
procedure the returned basket contains not the accumulation of the
traversal steps but only the nodes obtained during the last step. This
is rule is not compatible with the 'until no new nodes are found' end
iteration criteria.

* RuleSequence: this can concatenate the application of different rules
(it basically works like an UpdateRule that iterates over a chain of
rules instead of a single querybuilder instance).

* RuleSaveWalkers and RuleSetWalkers: rules that can be provided in a
chain of rules given to a RuleSequence to save a given state of the
current basket (Save) that can later be used to overwrite the content
of said working basket (Set). This is useful in the case where one might
need to do two operations 'in parallel' (i.e. on the same set of nodes)
instead of doing the second on the results of the first one.

Co-Authored-By: ramirezfranciscof <ramirezfranciscof@users.noreply.github.com>