Add feature traverse_graph and others
The function traverse_graph works as a simplified interface to interact
with the AGE that also removes the need to manually handle the basket
and the querybuilder instance:

 * The price to pay for hiding the basket is that this function can only be
   used with sets of nodes and links (so, no groups).

 * The price to pay for hiding the querybuilder is that complex traversal
   procedures can no longer be specified, the user simply defines which
   links can be traversed forwards and which backwards, and this criteria
   is then applied in every iteration (so one could not, in a single call,
   search only for all called calc nodes of the  called work nodes of an
   initial workflow node, as one will also obtain the calc nodes directly
   called by that initial workflow).

Besides the starting nodes (pks) and links, the user can also provide the
number of max iterations desired (which by default is None, which means
'until no new nodes are found') and a boolean that indicates if the links
(edges) should be returned.

Additionally, two other interfaces are included for ease of use when
deleting and exporting. These functions only take the starting set of
pks and the rules provided by the user (as 'rule_name_dir' = False/True)
and can automatically check if the rule is toggable, set defaults (using
aiida.common.links.GraphTraversalRules), and also parse the ruleset into
two lists with the links for forward and backward traversal. They will
return a dictionary containing the 'nodes' list, the 'links' list (if
this was requested, else this will contain `None`) and a dict with the
way in which all the rules were applied (using the following format:
'rule_name' = True/False).

Co-Authored-By: Leonid Kahle <leonid.kahle@epfl.ch>