Add traverse_graph / AGE as engine for node delete
The node deletion function now uses the get_nodes_delete function (with
the traverse_graph underlying interface using AGE as main engine) to
collect the extra nodes that are needed to keep a consistent provenance.
The procedure is not very different than the one that was initially
implemented so no significant performance improvement is expected, but
this is an important first step to homogenize graph traversal throughout
the whole code.