Consider 'AIIDA_TEST_PROFILE' in 'get_test_backend_name'.
Change the logic of 'get_test_backend_name' to check (if given)
the backend set in the configuration for the profile specified
in 'AIIDA_TEST_PROFILE'. If a backend is also specified in
'AIIDA_TEST_BACKEND', it checks that the two match, raising
ValueError otherwise. If neither is specified, fall back to the
default django backend.