Add the `verdi node repo dump` command (#3623)
This command can be used to easily dump the entire contents of a node's
repository to a folder on the local file system. To prevent accidentally
overwriting existing files, the command requires that the path to a non
existing directory be specified where the contents are to be stored. If
the path already exists, the command will abort.