Skip documenting the outline if it is None
If no outline is created in the 'define' method, 'get_outline'
will return None. Previously, this led to an error in the sphinx
directive. This is fixed by explicitly checking for None.

This error occured in an 'abstract base' workchain that does not
implement an outline itself, but is still documented.