Do not hijack the SqlAlchemy session factory in Django `QueryBuilder`
Since the `QueryBuilder` implementation uses SqlAlchemy to map the query
onto the models in order to generate the SQL to be sent to the database,
it requires a session, which is an :class:`sqlalchemy.orm.session.Session`
instance. The only purpose is for SqlAlchemy to be able to connect to the
database perform the query and retrieve the results. Even the Django
backend implementation will use SqlAlchemy for its `QueryBuilder` and so
also needs an SqlA session. It is important that we do not reuse the
scoped session factory in the SqlAlchemy implementation, because that
runs the risk of cross-talk once profiles can be switched dynamically in
a single python interpreter.

Therefore the Django implementation of the `QueryBuilder` should keep its
own SqlAlchemy engine and scoped session factory instances that are used
to provide the query builder with a session.