Update `prospector` to the latest version `v1.2.0`
This comes with new requirement of `pylint==2.4.4` which comes with a
few new warnings that have been addressed in the code base:

 * no-else-break
 * no-else-continue
 * self-assigning-variable
 * property-with-parameters
 * invalid-overridden-method
 * unnecessary-comprehension

One additional new warning has been added to the ignore list

 * import-outside-toplevel

PEP8 recommends to only have imports at the toplevel of a module and not
in function or class definitions. However, the changes required would be
significant and so we leave that for another time.