`pylint` will shut up one way or the other (#3696)
It keeps complaining about `requested_wallclock_time_seconds` being an
invalid name in `aiida.schedulers.plugins.slurm` however it is disabled
with an inline statement that apparently gets ignored. It could be that
another instance of this variable is triggering the warning and the
source file is just incorrect. So here we add the disable statement for
all instances of `requested_wallclock_time_seconds` and pray to god that
this time it works.