`CalcJob`: do not pause when exception thrown in the `presubmit`
The `presubmit` call of the `CalcJob` has recently been moved to the
`aiida.engine.processes.calcjobs.tasks.task_upload_job` which is wrapped
in the exponential backoff mechanism. The latter was introduced to
recover from transient problems such as connection problems during the
actual upload to a remote machine. However it should not catch exception
from the `presubmit` call which are not actually transient and thus not
automatically recoverable. In this case the process should simply
except.

Here we test this by mocking the presubmit to raise an exception and
check that it is bubbled up and the process does not end up in a paused
state. To prevent the test from blocking in the case the process gets
put in the paused state erroneously, we put a timeout on the test for
which we need the `pytest-timeout` plugin.