Remove `recreate_after_fork` trigger for SqlAlchemy scoped session
Presumably, the `multiprocessing.recreate_after_fork` hook on the
session creation of the SqlAlchemy was added because of the design of
the old daemon of the v0.* series. There the daemon was implemented
using `celery` which would spawn workers, by forking the main process.
Since the definition of forking process means "cloning a process with
the exact same state", the forked process would have the same session
which is undesirable. This is why the session has to be reset such that
an independent one has to be initialized.

With the new engine design, there is no forking anymore by the daemon
and it is also not supported. Instead, each daemon worker is launched as
a completely new process and so will not start with an existing session.

Finally, the `multiprocessing.recreate_after_fork` is not even
documented and according to `https://bugs.python.org/issue21372` this is
for a reason, because it is not intended for external use. Combined with
the fact that its cross-platform consistent operation is pulled into
question, it is better to not use it.