Update dependeny requirement `circus~=0.16.1` (#3703)
This new release of `circus` makes it compatible with `pyzmq>=17`. This
in turn allows us to unpin the `ipython` requirement which required
these versions of `pyzmq`. Note that implicitly the upper limit is
`7.10` for `ipython` because there python 3.5 support is dropped which
we support for another 9 months.

Note that we specify `0.16.1` because the uploaded wrong tarball was
uploaded for `0.16.0` making installation from source fail.