Add `exit_codes` argument to `register_process_handler`
The `exit_codes` argument takes a single or list of `ExitCode`
instances. If defined, the handler will return `None` if the exit
code set on the `node` does not appear in the `exit_codes`. This is
useful to have a handler called only when the process failed with a
specific (set of) exit code(s).