Move all tests from `aiida.backends.tests` to top level module
This will prevent them from being included in the distribution that is
uploaded to PyPI making the download a lot quicker.

Note that this commit only moves the files without touching them. This
means that the tests won't actually run since, for example, various
imports are broken. For clarity these changes will be done in a
following commit to not drown out the important changes between files
that were just moved around.

Note that the `django` and `sqlachemy` modules in `tests/backend`, which
contain backend specific tests, are prefixed with `aiida_`, otherwise
imports using either `from django` or `from sqlalchemy` will target
incorrectly those modules, instead of the actual libraries.