`BaseRestartWorkChain`: add `process_handlers` input namespace
This namespace allows one to register a process handler dynamically
through the inputs of a `BaseRestartWorkChain`. The `process_handlers`
namespace is a mapping of `ProcessHandler` named tuples. By using the
named tuple as type, the user can specify the same arguments as they
would when using the `register_process_handler` decorator to register a
process handler, such as the `priority` and `exit_codes` filter.

Another advantage of this is that this makes handling and validation of
these objects cleaner than just free-form tuples. The `ProcessHandler`
instances passed through the inputs are actually registered on the work
chain class, by calling the `register_process_handler` decorator
dynamically in the `BaseRestartWorkChain._load_handlers`. This approach
has the advantage that handlers registered dynamically through the
inputs behave exactly as if they would have been registered using the
decorator.

In cases where `ProcessHandlers` are passed through the inputs, they
will be serialized in the checkpoint. Since they contain an actual
function, the 'method' member of the `Processhandler`, a representer
and constructor for functions is added to `aiida.orm.utils.serialize`.