Fix `aiida.common.folders.Folder` unit test for utf-8 encoding
The test was calling `insert_path` from a filepath into itself. When
moving to python 3.8 this was triggering a weird recursive condition in
the underlying `shutil.copytree` call that kept concatenating the
`šaltinis/destination` sub string to the target filepath. This would
eventually throw an error that the filepath was too long. The origin for
this occurring all of a sudden in python 3.8 is unsure, but most likely
the recursive copy in the test was a typo and not intended.