Improve `restapi.common.identifiers.get_node_namespace` efficiency
This function serves to build a full hierarchical namespace of all
existing node types in the database. It is used by the REST API to
populate the side bar accordion with all the available node types, which
facilitates easy filtering. However, for big databases this function is
very slow. The main cost is the query for node/process type tuples:

    QueryBuilder().append(Node, project=['node_type', 'process_type'])

The function really only needs the unique tuples, which is currently
done in python. This can be done straight on the SQL level by using:

    builder.distinct()

This prevents having to load the tuples for all nodes, which saves a lot
of computing time for big databases. By adding the `distinct` clause,
the query for a database of approximately 3 million nodes, was reduced
by a factor of 10, from 50 to roughly 5 seconds.