Add `provenance_exclude_list` attribute to `CalcInfo` data structure (#3720)
This new attribute takes a flat list of relative filepaths, which
correspond to files in the `folder` sandbox passed to the
`prepare_for_submission` call of the `CalcJob`, that should not be
copied to the repository of the `CalcJobNode`. This functionality is
useful to avoid the content of input files, that should be copied to the
working directory of the calculation, to also be stored permanently in
the file repository. Example use cases are for very large input files or
files whose content is proprietary. Both use cases could already be
implemented using the `local_copy_list` but only in the case of files of
an input node in its entirety. The syntax of the `local_copy_list` does
not support the exclusion of arbitrary files that are written by the
calculation plugin to the sandbox folder.

Before the addition of this new feature, the contents of the sandbox
folder were added to the repository of the calculation node simply by
moving the contents of the sandbox entirely to the repository. This was
changed to an explicit loop over the contents and only copying those
files that do not appear in the `provenance_exclude_list` list.

The advantage of recursively looping over the contents of the sandbox
folder and *copying* the contents to the repository as long as it is not
part of `provenance_exclude_list`, over deleting those excluded files
from the sandbox before *moving* the remaining content to the
repository, is that in the former there is a better guarantee that the
excluded files do not accidentally end up in the repository due to an
unnoticed problem in the deletion from the sandbox.

The moving method is of course a lot more efficient then copying files
one by one. However, this moving approach is only possible now that the
repository is still implemented on the same filesystem as the sandbox.
Once the new repository interface is fully implemented, where non file
system repositories are also possible, moving the sandbox folder to the
repository will no longer be possible anyway, so it is acceptable to
already make this change now, since it will have to be done at some
point anyway.