select python version for conda CI test
It turns out that when creating a conda environment from an
`environment.yml` file, it is not possible to specify or override the
python version of the environment from the command line.

I.e. it turns out that

   conda env create -f environment.yml -n test-environment python=3.7

actually sets up a python 3.6 environment.

Until this functionality becomes available, I suggest we explicitly set
the python version it in the `environment.yml` file.