Conda: define explicit python requirement in `environment.yml` (#3758)
It turns out that when creating a conda environment from an
`environment.yml` file, it is not possible to specify or override the
python version of the environment from the command line.

I.e. it turns out that:

   conda env create -f environment.yml -n test-environment python=3.7

actually sets up a python 3.6 environment. This was because the
dependency `plumpy` was not yet marked as being compatible with
python 3.7. This has since been fixed and we now explicitly set the
python version in the `environment.yml` file.