Fix bugs in _store_from_cache and repository.erase
The _store_from_cache needed to erase the content of its sandbox
folder before copying over the content of the cache source.
Currently, the existing contents are mixed with with the contents
to be copied in.
In fixing this, we discovered an issue in repository.erase, where
an unstored node would try erasing the (inexistent) folder in the
permanent repository, instead of the SandboxFolder.