Add regression test for bug in store_from_cache
The regression test creates a Data node with a file in a directory
in the repository. After storing, it creates a clone, and stores
it when caching is enabled. Finally, the hashes of the original
and clone are compared.
Checked that this test fails when the store_from_cache fix is not
present.